2nd version of nav
I've created the 2nd version of the nav bar, which is going to replace the old one. Its been built mobile first, and I'm currently working on styles for a screen width of 1000px+.  I will need to create several other media queries to adjust for the logo so that the links and logo don't collide. I also need to make a fix for mobile so that the logo and hamburger menu icon don't collide either.