Footer
Finished the footer for mobile and desktop. Might make changes later but for the time being it's alright. I've also started working on a button to click to take you down to a section that will be about Eagle Airline Express, and maybe go into the Eagle Points Reward system, but that section has yet to be made